+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.0.2] - 2020-03-24
 ### Fixed
-- OperitionId to Psr normalizer class name 
+- OperitionId to Psr normalizer class name
 
 ## [0.0.1] - 2020-03-24
 ### Added