+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Data Transfer Object classes generation from swagger spec
+
+## [0.0.3] - 2020-04-27
+### Added
 - Neomerx Encoder implementation aware about Doctrine Proxies
 - DTO super classes
 
@@ -18,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - FastRoute Dispatcher generation from swagger operationIds
 
-[Unreleased]: https://github.com/FreeElephants/json-api-php-toolkit/compare/0.0.2...HEAD
+[Unreleased]: https://github.com/FreeElephants/json-api-php-toolkit/compare/0.0.3...HEAD
+[0.0.3]: https://github.com/FreeElephants/json-api-php-toolkit/compare/0.0.2...0.0.3
 [0.0.2]: https://github.com/FreeElephants/json-api-php-toolkit/compare/0.0.1...0.0.2
 [0.0.1]: https://github.com/FreeElephants/json-api-php-toolkit/releases/tag/0.0.1